backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pg8000
langchain
langchain-community
langchain-groq
sentence-transformers
pydantic
python-dotenv
requests
aiofiles
python-multipart
pgvector
streamlit
pandas
numpy
